[Reply] Replace with Kotlin Flows and new Architecture
